moved types that are OGM related to separate sub namespace
