+++ docs/node.md
@@ -29,8 +29,8 @@ Nodes are entities which contain the following attributes:
 ## Examples
 
 ```php
-use Syndesi\CypherDataStructures\Type\NodeLabel;
-use Syndesi\CypherDataStructures\Type\Node;
+use Syndesi\CypherDataStructures\Type\OGM\NodeLabel;
+use Syndesi\CypherDataStructures\Type\OGM\Node;
 use Syndesi\CypherDataStructures\Type\PropertyName;
 
 $node = new Node();
