+++ src/Type/OGM/NodeLabelStorage.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Syndesi\CypherDataStructures\Type;
+namespace Syndesi\CypherDataStructures\Type\OGM;
 
 use SplObjectStorage;
 use Syndesi\CypherDataStructures\Contract\NodeLabelInterface;
